refactor(area): Socials takes one typed contacts object #1176
extractContacts moves to $lib/area/contacts, shared by the SSR bundle
and both client consumers. Socials' interface collapses from 22 named
props to one AreaContacts object (its internals and sanitization keep
their per-key names); the communities-map popup and CommunityCard drop
their own 22-line contact:* fan-outs for the same helper.
